// src/shared/api/models/getApiStatisticsBundleOffers200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiStatisticsBundleOffers200SoldBundleOffersItem } from './getApiStatisticsBundleOffers200SoldBundleOffersItem';

/**
 * Successfully retrieved bundle offer statistics
 */
export type GetApiStatisticsBundleOffers200 = {
    totalAmount: number;
    totalOriginalPunches: number;
    totalResiduePunches: number;
    soldBundleOffers: GetApiStatisticsBundleOffers200SoldBundleOffersItem[];
};