// src/shared/api/models/getApiStatisticsBundleOffersGrouping.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiStatisticsBundleOffersGrouping =
    (typeof GetApiStatisticsBundleOffersGrouping)[keyof typeof GetApiStatisticsBundleOffersGrouping];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiStatisticsBundleOffersGrouping = {
    true: 'true',
    false: 'false',
} as const;